requirements.txt
pytorch
argparse
sklearn
wandb
matplotlib
seaborn
SciencePlots